+++ study/src/main/java/com/umc/study/domain/Mission.java
@@ -1,4 +1,4 @@
-package com.umc.study;
+package com.umc.study.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
