+++ study/src/main/java/com/umc/study/domain/Store.java
@@ -1,4 +1,4 @@
-package com.umc.study;
+package com.umc.study.domain;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
